23-7-2019 - 09:45 Config and useri/p

+++ Machine_BubblePoint_4.3/src/userinput/NSelectproject1.java
@@ -865,7 +865,7 @@ public class NSelectproject1 implements Initializable {
 		{
 			try
 			{
-				int re=Integer.parseInt(txtthiknes.getText());
+				Float re=Float.parseFloat(txtthiknes.getText());
 				
 			}
 			catch(Exception es)
